+++ src/Components/AuthenticationView.js
@@ -1,0 +1,18 @@
+import React from 'react';
+import PropTypes from 'prop-types';
+import { Button } from '@deskpro/apps-components';
+
+export const AuthenticationView = () => {
+
+  return (
+    <div>
+      <p>
+        This application has not been properly installed.
+      </p>
+
+      <p>
+        Contact the admin to finalize the set up.
+      </p>
+    </div>
+  );
+};