+++ src/Mailchimp/index.js
@@ -1,0 +1,6 @@
+
+export {FetchClient as MailchimpFetchClient} from './APIClient'
+export {fetchSubscriberInfo, fetchLists, updateListSubscriptions, determineMembershipDetails} from './Services'
+export {ApiCredentials, MailchimpAuthcInfo} from './Authentication'
+
+export {AuthenticationError as MailchimpAuthenticationError} from './Errors'