+++ src/main/javascript/Components/index.js
@@ -1,7 +1,0 @@
-
-export { MemberActivityList } from './CampaignsSection'
-export { SubscriptionsList } from './SubscriptionsSection'
-export { UserCard } from './UserCardSection'
-export { HomeView } from './HomeView'
-export { AuthenticationView } from './AuthenticationView'
-export { OauthConnectionView } from './OauthConnectionView'